requirements.txt
huggingface-hub==0.26.2
matplotlib==3.7.5
numpy==1.24.4
opencv-python==4.9.0.80
opencv-python-headless==4.10.0.84
pandas==2.0.3
pillow==10.4.0
protobuf==5.28.3
psutil==6.1.0
pybboxes==0.1.6
pycocotools==2.0.7
pynvml==11.5.3
pyparsing==3.1.4
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytools==2024.1.14
pytz==2024.2
PyYAML==6.0.2
scikit-learn==1.3.2
six==1.16.0
torch==2.4.1
torchmetrics==0.10.0
torchvision==0.19.1
tornado==6.4.1


